app_project/backend/requirements.txt
fastapi
uvicorn
pyodbc
sqlalchemy
pydantic_settings
pymssql
jinja2
PyJWT==2.8.0
passlib==1.7.4
python-multipart
python-jose
email-validator
bcrypt==3.2.0